pyproject.toml
[project]
name = "langchain-tut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-groq>=0.3.5",
    "langgraph>=0.5.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
]
